# swagger.yaml
openapi: 3.0.0
info:
  title: Trip Booking Api Documentation
  version: 1.0.0
  description: Trip Api Document
servers:
  - url: http://localhost:5000
paths:
...
  /api/tours/tour/{id}:
    get:
      summary: Get a tour by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tour found
        '404':
          description: Tour not found

    patch:
      summary: Update a tour by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TourUpdate'
      responses:
        '200':
          description: Tour updated

    delete:
      summary: Hard delete a tour by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tour deleted
...
components:
  schemas:
...
    TourUpdate:
        type: object
        properties:
          title:
            type: string
            example: "Giza pyramids and Sphinx. “With camel ride”"
          price:
            type: number
            format: double 
            example: 199.99
          Description:
            type: string
            example: "Explore the Pyramids of Giza and the Great Sphinx on a half-day tour of Cairo with hotel transfers. Enjoy a short camel ride near the pyramids, immersing yourself in the spirit of ancient Egypt."
          FullDescription:
            type: string
            example: "Explore the Pyramids of Giza and the Great Sphinx on a half-day tour of Cairo with hotel transfers. Enjoy a short camel ride near the pyramids, immersing yourself in the spirit of ancient Egypt."
          Duration:
            type: number
            example: 3
          included:
            type: array
            items:
              type: string
            example: ["Hotel pickup and drop-off", "Transportation by air-conditioned vehicle", "Entry ticket"]
          excluded:
            type: array
            items:
              type: string
            example: ["Entry ticket to go inside any Pyramid", "Lunch expenses","Tips"]
          locations:
            type: array
            items:
              type: string
            example: ["Giza pyramids", "Sphinx"]
          highlight:
            type: array
            items:
              type: string
            example: ["Uncover ancient Egyptian history on a private tour of the Giza Pyramids", "Marvel at the Great Sphinx of Giza, a limestone statue of a reclining sphinx"]
          cloudImages:
            type: array
            items:
              type: string
              format: binary
            example:
              - "https://res.cloudinary.com/.../image1.jpg"
              - "https://res.cloudinary.com/.../image2.jpg"
          localImages:
            type: array
            items:
              type: string
              format: binary
            example:
              - "/uploads/image1.jpg"
              - "/uploads/image2.jpg"
          IsDeleted:
            type: boolean
            default: false